feat: add endpoint DELETE in class Ordem de Serviço

+++ src/main/java/com/devffl/babershop/repositories/OrdemServicoRepository.java
@@ -3,5 +3,5 @@ package com.devffl.babershop.repositories;
 import com.devffl.babershop.entities.OrdemServico;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrdemServicoRepository extends JpaRepository<OrdemServico, Integer> {
+public interface OrdemServicoRepository extends JpaRepository<OrdemServico, Long> {
 }
